Split runner shell profile tests
